+++ src/styles/EmailAction.css
@@ -1,0 +1,72 @@
+.email-action-1 {
+    display: flex;
+    justify-content: center;
+    align-items: flex-end; 
+    background-color: #f4f4f9;
+    font-family: Arial, sans-serif;
+    margin-left: auto;
+  }
+  
+  .email-content-1 {
+    background: #fff;
+    border-radius: 8px;
+    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+    width: 500px;
+    padding: 20px;
+  }
+  
+  .email-content--header-1 {
+    display: flex;
+    align-items: center;
+    border-bottom: 1px solid #ddd;
+    padding-bottom: 10px;
+    margin-bottom: 20px;
+  }
+  
+  .avatar1 {
+    width: 50px;
+    height: 50px;
+    background-color: #ccc;
+    border-radius: 50%;
+    margin-right: 15px;
+  }
+  
+  .sender-info1 h2 {
+    margin: 0;
+    font-size: 18px;
+  }
+  
+  .sender-info1 em {
+    color: #888;
+    font-size: 14px;
+  }
+  
+  .reply-body1 textarea {
+    width: 100%;
+    height: 100px;
+    border: 1px solid #ddd;
+    border-radius: 4px;
+    padding: 10px;
+    font-size: 14px;
+    margin-bottom: 20px;
+  }
+  
+  .email-actions {
+    display: flex;
+    justify-content: flex-end;
+  }
+  
+  .email-actions button {
+    background-color: #007bff;
+    color: #fff;
+    border: none;
+    border-radius: 4px;
+    padding: 10px 20px;
+    cursor: pointer;
+    font-size: 14px;
+  }
+  
+  .email-actions button:hover {
+    background-color: #0056b3;
+  }
+
