+++ src/styles/EmailView.css
@@ -1,0 +1,123 @@
+
+.email-toolbar {
+    padding: 0.5rem 2rem;
+  
+    display: grid;
+    grid-template-columns: auto 1fr auto;
+  
+    border-bottom: 1px solid #eee;
+  }
+  
+  .email-toolbar > * {
+    display: grid;
+    grid-auto-flow: column;
+    align-items: center;
+    gap: 1rem;
+  }
+  
+  .email-toolbar > ul {
+    list-style-type: none;
+  }
+  
+  .email-toolbar .icon {
+    width: 24px;
+    height: 24px;
+  }
+  
+  .email-toolbar button {
+    padding: 0.25rem 0.5rem 0.4rem;
+  
+    border-radius: 0.5rem;
+    border: 1px solid #a1a1a1;
+    background-color: transparent;
+  
+    font-weight: bold;
+  }
+  
+  .title {
+    padding: 1rem;
+  }
+  
+  .title h1 {
+    padding-left: 56px;
+  
+    font-weight: normal;
+  }
+  
+  .email-content--header {
+    padding: 0.5rem 1rem;
+  
+    display: grid;
+    grid-template-columns: auto 1fr auto auto;
+    gap: 1rem;
+  }
+  
+  .email-content--header .avatar {
+    width: 56px;
+    height: 56px;
+  
+    border-radius: 100%;
+    background-color: #eee;
+  }
+  
+  .email-info {
+    padding: 0.25rem 0;
+  }
+  
+  .email-info .sender-info h2 {
+    margin-right: 1rem;
+  
+    display: inline-block;
+  
+    font-size: 0.925rem;
+  }
+  
+  .email-info .user-info {
+    margin-top: 0.5rem;
+  }
+  
+  .email-info .user-info p {
+    font-size: 0.825rem;
+  }
+  
+  .email-action-icons ul {
+    padding: 0;
+    margin: 0;
+  
+    display: grid;
+    grid-auto-flow: column;
+    gap: 0.25rem;
+  
+    list-style-type: none;
+  }
+  
+  .email-action-icons .icon {
+    width: 20px;
+    height: 20px;
+  }
+  
+  .email-body .content {
+    max-width: 600px;
+    margin: 0 auto;
+  }
+  
+  .email-body .content img {
+    width: 100%;
+  }
+  
+  .email-actions {
+    padding: 1rem 0.5rem;
+  }
+  
+  .email-actions button {
+    padding: 0.5rem 1rem;
+    margin-right: 0.5rem;
+  
+    border-radius: 0.5rem;
+    border: 1px solid #a1a1a1;
+    background-color: transparent;
+  
+    font-weight: bold;
+  }
+
+
