+++ src/styles/Header.css
@@ -1,4 +1,4 @@
-.header {
+.app > .header {
     grid-area: header;
   
     display: grid;
@@ -6,27 +6,27 @@
   
     border-bottom: solid 1px #00000020;
   }
+  
 
-
-.header > .left-menu {
+.app > .header > .left-menu {
     display: grid;
     grid-auto-flow: column;
     align-items: center;
     justify-content: left;
   }
   
-.header > .left-menu .menu-icon {
+  .app > .header > .left-menu .menu-icon {
     width: 25px;
     height: 25px;
     margin: 10px;
   }
   
-.header > .search {
+  .app > .header > .search {
     display: grid;
     padding: 6px;
   }
-
-  .header > .search .search-bar {
+  
+  .app > .header > .search .search-bar {
     border: none;
     border-radius: 5px;
     background-color: rgb(240, 240, 240);
